refactor(entities): change column type from numeric to int for order field
The 'order' field in the Question entity was updated from 'numeric' to 'int' to better represent its data type and improve consistency with the expected integer values.

+++ src/survey/entities/question.entity.ts
@@ -19,7 +19,7 @@ export class Question {
 	})
 	type: QuestionType;
 
-	@Column({ type: 'numeric', default: 0 })
+	@Column({ type: 'int', default: 0 })
 	order: number;
 
 	@CreateDateColumn({ type: 'timestamptz' })
